fix: make user-metadata size guard compile on 32-bit platforms
The size guard compared an int length against math.MaxUint32, an untyped
int constant. On 32-bit platforms (e.g. GOOS=linux GOARCH=386) MaxUint32
overflows int, so the comparison and the fmt.Errorf argument both failed
to compile, regressing a build that previously worked. Cast the length to
uint64 rather than forcing the constant into an int; behaviour on 64-bit
is unchanged and the package now builds on every architecture.

+++ index_writer.go
@@ -55,8 +55,12 @@ type indexWriter struct {
 func newIndexWriter(path string, cfg *buildConfig, numBlocks uint32, algo blockBuilder) (*indexWriter, error) {
 	// Compute variable section sizes
 	userMetadataLen := len(cfg.userMetadata)
-	if userMetadataLen > math.MaxUint32 {
-		return nil, fmt.Errorf("user metadata size %d exceeds maximum of %d bytes", userMetadataLen, math.MaxUint32)
+	// Cast the length (not the constant) to uint64: math.MaxUint32 is an untyped
+	// int constant, so comparing it against an int operand overflows int and fails
+	// to compile on 32-bit platforms. uint64 holds MaxUint32 on every architecture
+	// and the comparison is identical on 64-bit.
+	if uint64(userMetadataLen) > math.MaxUint32 {
+		return nil, fmt.Errorf("user metadata size %d exceeds maximum of %d bytes", userMetadataLen, uint64(math.MaxUint32))
 	}
 	algoConfigLen := algo.GlobalConfigSize()
